requirements.txt
line-bot-sdk
bs4
flask
lask-cors
pymongo
datetime
pandas
SnowNLP
emoji
pyshorteners
scipy
cors
gunicorn
